requirements.txt
curl_cffi
eth-account
eth-utils
websockets
aiohttp
aiohttp-socks
tortoise-orm
loguru
rich
colorama
pyyaml
aiofiles